third_party/WebKit/LayoutTests/imported/wpt/offscreen-canvas/the-offscreen-canvas/size.attributes.parse.onlyspace-expected.txt

This is a testharness.js-based test.
FAIL Parsing of non-negative integers assert_equals: offscreenCanvas.width === 300 (got 100[number], expected 300[number]) expected 300 but got 100
Harness: the test ran to completion.
